feat: v0.0.8 stability & architecture overhaul
Closes #130 #131 #132 #134 #138 #139 #144 #137

Architecture:
- Split lib.rs (1471→352) and main.rs (1538→422) into focused modules
- Core: operations, maintenance, consolidate, error, types, import_export
- CLI: commands, init, output, bench

Validation (#132):
- Max content 10K chars, max 20 tags, max server payload 1MB

Binary integrity (#134):
- SHA256 checksums in release artifacts (CHECKSUMS.txt)
- verify-checksums subcommand

Schema versioning (#138):
- schema_version table + migration framework

Reliability (#139):
- Fix deadlock in check_contradiction (mutex re-acquire)

Error handling (#144):
- Error enum rewrite with sanitized user-friendly messages
- ~90 call sites migrated from raw rusqlite/usearch/ONNX errors
- Lock errors now include operation context
- 7 new unit tests for sanitization

Python wrapper (#137):
- 7→21 methods covering all CLI commands
- Namespace support on all methods
- Full type hints + Google-style docstrings

Benchmark (#49):
- memory-bench binary for library-level timing

README:
- Badge v0.0.7, Design Philosophy section
- Performance section with actual benchmark data
- Removed knowledge graph from roadmap

Tests: 116 pass (+20 from v0.0.7)
