+++ Cargo.toml
@@ -22,3 +22,7 @@ lto = true          # Link-time optimization
 codegen-units = 1   # Single codegen unit for better LTO
 strip = true        # Strip debug symbols
 panic = "abort"     # Smaller panic handler
+
+[profile.bench]
+opt-level = 3
+lto = true
