+++ crates/uteke-cli/Cargo.toml
@@ -10,6 +10,10 @@ description = "CLI for uteke — persistent memory for AI agents"
 name = "uteke"
 path = "src/main.rs"
 
+[[bin]]
+name = "memory-bench"
+path = "src/bench.rs"
+
 [dependencies]
 uteke-core = { path = "../uteke-core" }
 clap = { version = "4", features = ["derive"] }
